Uniformization before merge on main (#25)
* uniformization before merge on main

* fix routing links
