+++ TeamCreator/src/app/Views/auth/auth.component.html
@@ -1,0 +1,8 @@
+<form [formGroup]="authForm" (ngSubmit)="onSubmit()">
+
+    <label for="username">Login : </label><br>
+    <input required id="username" placeholder="Username" type="text" formControlName="name" name="name">
+  
+    <button type="submit">Submit</button>
+  </form>
+
