+++ TeamCreator/src/app/Views/auth/auth.component.ts
@@ -1,13 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import {FormControl, FormGroup} from "@angular/forms";
 import {Router} from "@angular/router";
-import {User} from "../user";
+import { User } from 'src/app/Models/user';
 
 @Component({
   selector: 'app-auth',
   templateUrl: './auth.component.html',
   styleUrls: ['./auth.component.css']
 })
+
 export class AuthComponent implements OnInit {
   authForm = new FormGroup({
     name: new FormControl(''),
