+++ TeamCreator/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import {HomeComponent} from "./home/home.component";
-import {AuthComponent} from "./auth/auth.component";
+import { AuthComponent } from './Views/auth/auth.component';
+import { HomeComponent } from './Views/home/home.component';
 
 const routes: Routes = [
   { path: 'home', component: HomeComponent },
