--- TeamCreator/src/app/auth/auth.component.css
@@ -1,1 +1,0 @@
-@import url('https://unpkg.com/bootstrap@3.3.7/dist/css/bootstrap.min.css');
