--- TeamCreator/src/app/auth/auth.component.html
@@ -1,8 +1,0 @@
-<form [formGroup]="authForm" (ngSubmit)="onSubmit()">
-
-  <label for="username">Login : </label><br>
-  <input required id="username" placeholder="Username" type="text" formControlName="name" name="name">
-
-  <button type="submit">Submit</button>
-</form>
-
