--- TeamCreator/src/app/home/home.component.css
@@ -1,52 +1,0 @@
-/* usersComponent's private CSS styles */
-.users {
-  margin: 0 0 2em 0;
-  list-style-type: none;
-  padding: 0;
-  width: 15em;
-}
-.users li {
-  cursor: pointer;
-  position: relative;
-  left: 0;
-  background-color: #EEE;
-  margin: .5em;
-  padding: .3em 0;
-  height: 1.6em;
-  border-radius: 4px;
-}
-.users li:hover {
-  color: #2c3a41;
-  background-color: #e6e6e6;
-  left: .1em;
-}
-.users li.selected {
-  background-color: black;
-  color: white;
-}
-.users li.selected:hover {
-  background-color: #505050;
-  color: white;
-}
-.users li.selected:active {
-  background-color: black;
-  color: white;
-}
-.users .badge {
-  display: inline-block;
-  font-size: small;
-  color: white;
-  padding: 0.8em 0.7em 0 0.7em;
-  background-color:#405061;
-  line-height: 1em;
-  position: relative;
-  left: -1px;
-  top: -4px;
-  height: 1.8em;
-  margin-right: .8em;
-  border-radius: 4px 0 0 4px;
-}
-
-input {
-  padding: .5rem;
-}
