--- TeamCreator/src/app/user.spec.ts
@@ -1,7 +1,0 @@
-import { User } from './user';
-
-describe('User', () => {
-  it('should create an instance', () => {
-    expect(new User()).toBeTruthy();
-  });
-});
